+++ lib/features/setting/bottom_sheets/components/bottom_sheet_base.dart
@@ -1,0 +1,69 @@
+// features/setting/components/bottom_sheet_base.dart
+import 'package:flutter/material.dart';
+
+class BottomSheetBase extends StatelessWidget {
+  final String title;
+  final Widget content;
+
+  const BottomSheetBase({
+    super.key,
+    required this.title,
+    required this.content,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    final colorScheme = theme.colorScheme;
+
+    return Container(
+      decoration: BoxDecoration(
+        color: colorScheme.surface,
+        borderRadius: const BorderRadius.vertical(top: Radius.circular(20)),
+      ),
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          _buildDragHandle(colorScheme),
+          const SizedBox(height: 20),
+          _buildHeader(context, title, theme, colorScheme),
+          const SizedBox(height: 20),
+          Flexible(
+            child: SingleChildScrollView(
+              physics: const BouncingScrollPhysics(),
+              padding: const EdgeInsets.symmetric(horizontal: 20),
+              child: content,
+            ),
+          ),
+          const SizedBox(height: 20),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildDragHandle(ColorScheme colorScheme) {
+    return Container(
+      margin: const EdgeInsets.only(top: 8),
+      width: 40,
+      height: 4,
+      decoration: BoxDecoration(
+        color: colorScheme.outline.withOpacity(0.3),
+        borderRadius: BorderRadius.circular(2),
+      ),
+    );
+  }
+
+  Widget _buildHeader(BuildContext context, String title, ThemeData theme,
+      ColorScheme colorScheme) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 20),
+      child: Text(
+        title,
+        style: theme.textTheme.headlineSmall?.copyWith(
+          fontWeight: FontWeight.w700,
+          color: colorScheme.primary,
+        ),
+      ),
+    );
+  }
+}
