+++ lib/features/setting/bottom_sheets/components/setting_section_header.dart
@@ -1,0 +1,44 @@
+// features/setting/components/setting_section_header.dart
+import 'package:flutter/material.dart';
+
+class SettingSectionHeader extends StatelessWidget {
+  final String title;
+  final IconData icon;
+
+  const SettingSectionHeader({
+    super.key,
+    required this.title,
+    required this.icon,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    final colorScheme = theme.colorScheme;
+
+    return Row(
+      children: [
+        Container(
+          padding: const EdgeInsets.all(8),
+          decoration: BoxDecoration(
+            color: colorScheme.primary.withOpacity(0.1),
+            borderRadius: BorderRadius.circular(8),
+          ),
+          child: Icon(
+            icon,
+            size: 18,
+            color: colorScheme.primary,
+          ),
+        ),
+        const SizedBox(width: 12),
+        Text(
+          title,
+          style: theme.textTheme.titleMedium?.copyWith(
+            fontWeight: FontWeight.w700,
+            color: colorScheme.primary,
+          ),
+        ),
+      ],
+    );
+  }
+}
