+++ lib/features/setting/section/app_info/app_info_page.dart
