+++ lib/features/setting/section/note_section/notes_setting_section.dart
@@ -9,7 +9,7 @@ import 'package:page_transition/page_transition.dart';
 import 'package:provider/provider.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:msbridge/core/provider/share_link_provider.dart';
-import 'package:msbridge/features/setting/pages/shared_notes_page.dart';
+import 'package:msbridge/features/setting/section/note_section/shared_notes_page.dart';
 import 'package:msbridge/core/repo/share_repo.dart';
 import 'package:msbridge/core/services/backup_service.dart';
 import 'package:msbridge/core/provider/sync_settings_provider.dart';
