+++ lib/features/setting/utils/time_formatter.dart
@@ -1,0 +1,17 @@
+// features/setting/utils/time_formatter.dart
+class TimeFormatter {
+  static String formatTimeAgo(DateTime dateTime) {
+    final now = DateTime.now();
+    final difference = now.difference(dateTime);
+
+    if (difference.inDays > 0) {
+      return "${difference.inDays}d ago";
+    } else if (difference.inHours > 0) {
+      return "${difference.inHours}h ago";
+    } else if (difference.inMinutes > 0) {
+      return "${difference.inMinutes}m ago";
+    } else {
+      return "Just now";
+    }
+  }
+}
