+++ pubspec.yaml
@@ -21,6 +21,7 @@ dependencies:
   line_icons: ^2.0.3
   intl: ^0.19.0
   google_fonts: ^6.2.1
+  webview_flutter: ^4.8.0
   flutter_highlight: ^0.7.0
   flutter_keyboard_visibility: ^6.0.0
   flutter_colorpicker: ^1.1.0
@@ -65,7 +66,6 @@ dependencies:
   flutter_local_notifications: ^16.3.2
   timezone: ^0.9.2
   markdown_widget: ^2.3.2+8
-  webview_flutter: ^4.8.0
 
 dev_dependencies:
   flutter_test:
